+++ controllers/options/check_user_rule.js
