+++ controllers/options/create_new.js
@@ -1,11 +1,11 @@
-const Store = require('../db_store/store')
-const db = require('../db_store/db')
-const { createSHHKey } = require('../ssh/ssh_controller');
+const Store = require('../../db_store/store')
+const db = require('../../db_store/db')
+const { createSHHKey } = require('../../ssh/ssh_controller');
 const { 
     USER_ALREADY_EXISTS,
     USERNAME_EMPTY,
     EMAIL_EMPTY
-} = require('../constants/global')
+} = require('../../constants/global')
 
 
 async function createNewUser(objc) {
