+++ controllers/options/delete_user.js
