+++ controllers/options/get_user_info.js
