+++ controllers/options/switch_user.js
