+++ controllers/options/update_user.js
